Chore: rename Timer to WindowTimer
