+++ test/utilities/WindowTimer.js
